shuffle a few more things into the cli package
