+++ rats-devtools/src/python/rats/ci/_plugin.py
@@ -2,10 +2,10 @@ from collections.abc import Callable
 
 import click
 
-from rats import apps
-
 # pyright seems to struggle with this namespace package
 # https://github.com/microsoft/pyright/issues/2882
+from rats import apps as apps
+from rats import cli as cli
 from rats import devtools as devtools
 
 from ._commands import RatsCiCommands
@@ -31,9 +31,9 @@ class RatsCiPlugin(apps.AnnotatedContainer):
         def get(name: str) -> Callable[[], click.Command]:
             return lambda: self._app.get(PluginServices.command(name))
 
-        return devtools.LazyClickGroup(
+        return cli.DeferredCommandGroup(
             name="ci",
-            provider=devtools.CommandProvider(
+            provider=cli.CommandProvider(
                 commands={name: get(name) for name in cmds},
             ),
         )
