+++ rats-devtools/src/python/rats/devtools/_groups.py
@@ -1,71 +1,33 @@
-from abc import abstractmethod
-from collections.abc import Callable, Iterator, Mapping
-from typing import Protocol
+from collections.abc import Iterator
 
 import click
 
-from rats import apps
+# pyright seems to struggle with this namespace package
+# https://github.com/microsoft/pyright/issues/2882
+from rats import apps as apps
+from rats import cli as cli
 
 from ._plugins import PluginRunner
 
 
-class CommandGroupPlugin(Protocol):
-    @abstractmethod
-    def on_group_registration(self, group: click.Group) -> None:
-        pass
-
-
 class CommandGroup(apps.Executable):
-    _plugins: PluginRunner[CommandGroupPlugin]
+    _plugins: PluginRunner[cli.CommandGroupPlugin]
 
-    def __init__(self, plugins: PluginRunner[CommandGroupPlugin]) -> None:
+    def __init__(self, plugins: PluginRunner[cli.CommandGroupPlugin]) -> None:
         self._plugins = plugins
 
     def execute(self) -> None:
         cli = click.Group()
-        self._plugins.apply(lambda plugin: plugin.on_group_registration(cli))
+        self._plugins.apply(lambda plugin: plugin.on_group_open(cli))
         cli()
 
 
-class CommandProvider:
-    _commands: Mapping[str, Callable[[], click.Command]]
-
-    def __init__(
-        self,
-        commands: Mapping[str, Callable[[], click.Command]],
-    ) -> None:
-        self._commands = commands
-
-    def list(self) -> frozenset[str]:
-        return frozenset(self._commands.keys())
-
-    def get(self, name: str) -> click.Command:
-        if name not in self._commands:
-            raise ValueError(f"Command {name} not found")
-
-        return self._commands[name]()
-
-
-class LazyClickGroup(click.Group):
-    _provider: CommandProvider
-
-    def __init__(self, name: str, provider: CommandProvider) -> None:
-        super().__init__(name=name)
-        self._provider = provider
-
-    def get_command(self, ctx: click.Context, cmd_name: str) -> click.Command | None:
-        return self._provider.get(cmd_name)
-
-    def list_commands(self, ctx: click.Context) -> list[str]:
-        return list(self._provider.list())
-
-
-class GroupCommands(CommandGroupPlugin):
+class GroupCommands(cli.CommandGroupPlugin):
     _commands: Iterator[click.Command]
 
     def __init__(self, commands: Iterator[click.Command]) -> None:
         self._commands = commands
 
-    def on_group_registration(self, group: click.Group) -> None:
+    def on_group_open(self, group: click.Group) -> None:
         for command in self._commands:
             group.add_command(command)
