+++ rats-devtools/src/python/rats/docs/_commands_annotated.py
