+++ rats-devtools/src/python/rats/docs/_plugin.py
@@ -2,7 +2,11 @@ from collections.abc import Callable
 
 import click
 
-from rats import apps, devtools  # type: ignore[reportAttributeAccessIssue]
+# pyright seems to struggle with this namespace package
+# https://github.com/microsoft/pyright/issues/2882
+from rats import apps as apps
+from rats import cli as cli
+from rats import devtools as devtools
 
 from ._commands import RatsCiCommands
 from ._ids import PluginServices
@@ -26,9 +30,9 @@ class RatsDocsPlugin(apps.AnnotatedContainer):
         def get(name: str) -> Callable[[], click.Command]:
             return lambda: self._app.get(PluginServices.command(name))
 
-        return devtools.LazyClickGroup(
+        return cli.DeferredCommandGroup(
             name="docs",
-            provider=devtools.CommandProvider(
+            provider=cli.CommandProvider(
                 commands={name: get(name) for name in cmds},
             ),
         )
